refactor: simplify withdraw logic and add separate logic for ata creation txns
