+++ src/pages/api/wallet/sign-transaction.ts
@@ -12,8 +12,6 @@ import { withAuth } from '@/features/auth/utils/withAuth';
 const TOKEN_PROGRAM_ID = 'TokenkegQfeZyiNwAJbNbGKPFXCWuBvf9Ss623VQ5DA';
 const COMPUTE_BUDGET_ID = 'ComputeBudget111111111111111111111111111111';
 const SYSTEM_PROGRAM_TRANSFER_ID = '11111111111111111111111111111111';
-const ASSOCIATED_TOKEN_PROGRAM_ID =
-  'ATokenGPvbdGVxr1b2hvZbsiqW5xWH25efTNsLJA8knL';
 
 const TOKEN_PROGRAM_TRANSFER_INSTRUCTION = 3;
 const SYSTEM_PROGRAM_TRANSFER_INSTRUCTION = 2;
@@ -23,7 +21,6 @@ const ALLOWED_PROGRAM_IDS = [
   TOKEN_PROGRAM_ID,
   COMPUTE_BUDGET_ID,
   SYSTEM_PROGRAM_TRANSFER_ID,
-  ASSOCIATED_TOKEN_PROGRAM_ID,
 ];
 
 async function validateTransaction(
